Add os flag check

+++ src/api/utils/test_base.py
@@ -5,15 +5,18 @@ import unittest
 from api.utils.factory import create_app
 from api.utils.database import db
 from api.utils.config import TestingConfig
+import os
 
 
 class BaseTestCase(unittest.TestCase):
     """A base test case"""
     def setUp(self):
-        app = create_app(TestingConfig)
-        app.app_context().push()
-        self.app = app.test_client()
+        if os.environ.get('WORK_ENV') == 'PROD':
+            app = create_app(TestingConfig)
+            app.app_context().push()
+            self.app = app.test_client()
 
     def tearDown(self):
-        db.session.close_all()
-        db.drop_all()
+        if os.environ.get('WORK_ENV') == 'PROD':
+            db.session.close_all()
+            db.drop_all()
